Add and fix test; no longer use kluent
I don't get why kluent doesn't show <Click to see difference> anymore: it worked before.

Commented on this ticket: https://github.com/MarkusAmshove/Kluent/issues/228
I also tried kotest assertions, but that also did not work: https://github.com/kotest/kotest/issues/4500

There is something very strange going on. The click to see the difference is a lifesaver, I switched to kotlin-test for this. Still: I don't like that: it's really hard to guess which one is the expected and which one is the actual, so I might switch again later.