Upload the real Google Maps key
Yes, this is the real Google Maps API key. This prevents me forgetting to fill it when making a release. This is fine: it is limited to Android apps and only when using with my debug / the upload keystore.

The hash of the certificates can be found by running the gradle command `signingreport`.