Greatly improve coordinate loading
We now load the coordinates and the locations from the backend at the same time
When the locations are loaded, show them based on the real coordinates when these are ready, otherwise, it first shows them based on the last known coordinates with isRefreshing=true.
Added a 5ms grace period for the locations to load normally. I can imagine this is enough when you just loaded your GPS location and prevents existing tests from breaking.