Also consider the max capacity from the history
This way, it's not too bad if max_2017_2023.csv is outdated, we'll at least always have the max capacity of the last week.

Doing this in the backend would still be better, as this would enable us to automatically get the max capacity of the last 2 months, but that's a lot more work than these 2 lines of code.