Use the maxCapacity the newly updated backend now returns
This greatly simplifies things: I now no longer have to update a CSV and release a new version of the app whenever there is more capacity somewhere. It also simplifies the code quite a bit.

Lastly, the new backend returns the max capacity of this month plus the 2 months before it, not the absolute max ever. I think that's better: there are some rare outliers (like when there's a massive storm) which bump the max capacity quite a bit. Also, I assume the max capacity is a bit seasonal.

Having too big of a max capacity can be a bad thing: when it's too much, it will show that only a tiny percentage of bikes are available because that's compared to the max capacity. This is not very realistic though when this max capacity hasn't occurred in over 3 months.

This change mitigates this and will show a more optimistic (and dare I say realistic) percentage in the details screen.