Only show the tablet layout starting at 800 width
It looks fine on my 800 dp phone on landscape, but terrible on a Pixel 7 portrait (at 600 dp).

There's an argument to be made to make the cut-off point at 840 dp, as that's where the expanded class cut-off is at, but it looks quite good on my 800 dp phone on landscape (and I suspect many others like it), so why not?

I suspect this is the reason why comparing classes is deprecated in favor of this method.